CustomDM - Food system revamp: Food no longer works like potions. This finally differentiates Cooking and Alchemy(note that stamina potions are craftable in CustomDM).
Instead of instantly recovering vitals food adds to an extra regeneration pool.

This pool is used to provide extra regeneration every time a regeneration event would occur, meaning it is only used when vitals are not full.

A full stomach should provide enough for about 2 hours of extra regeneration(if the vitals were always not at maximum so in practice it will last longer).